+++ app/src/main/java/com/codesoom/assignment/controllers/task/TaskUpdateController.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.controllers.task;
 
 import com.codesoom.assignment.domains.Task;
 import com.codesoom.assignment.domains.TaskDto;
@@ -14,9 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 /**
  * 할 일 수정 요청과 매핑됩니다.
  */
-@CrossOrigin
-@RequestMapping("/tasks")
-@RestController
+@TaskController
 public class TaskUpdateController {
 
     private final TaskUpdateService service;
